+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0-3.0 - UNRELEASED
+### Changed
+
+* Added support for own `KV` and `Value` implementations
+* Streamlined the formatting syntax for `log!` and friends; **BREAKING**
+* Lazy values need explicit `FnValue` wrapper
+
 ## 2.0.0-2.2 - 2017-03-19
 ### Fixes